Started to remove utility file and modularizing in classes for pipeline run and tests
