+++ src/helpers/utility.py
@@ -34,15 +34,6 @@ def decide_format(args: argparse.Namespace) -> argparse.Namespace:
     return args
 
 
-def check_index(path_index: str):
-    if Path(path_index).is_file():
-        pass
-    else:
-        exit(f"No index file found on {path_index}")
-
-    return
-
-
 def build_directory(args: argparse.Namespace, curr_time: str) -> argparse.Namespace:
     run(["mkdir", "-p", "results_" + curr_time + "/1_quality_control"])
     run(["mkdir", "-p", "results_" + curr_time + "/2_trimmed_output"])
